worker state
