+++ tests/test_worker_state.py
@@ -8,7 +8,6 @@ from pytraj.parallel import _worker_state
 from pytraj.tools import concat_dict
 
 
-@unittest.skip('parallel State: WIP')
 class TestWorkerState(unittest.TestCase):
     '''temp test. name will be changed.
     '''
@@ -30,22 +29,22 @@ class TestWorkerState(unittest.TestCase):
 
             lines = ['angle :3 :7 :9', 'distance :2 :10']
 
-            for n_cores in [2, 3, 4, 5, 6, 7, 8]:
+            for n_cores in [2, 3]:
                 data_list = [_worker_state(rank, n_cores, traj, lines)
                              for rank in range(n_cores)]
-                final_data = concat_dict(data_list)
-                aa_eq(final_data['Ang_00001'], saved_angle)
-                aa_eq(final_data['Dis_00002'], saved_dist)
+                final_data = concat_dict([x[1] for x in data_list])
+                aa_eq(final_data['Ang_00002'], saved_angle)
+                aa_eq(final_data['Dis_00003'], saved_dist)
 
     def test_multiple_cores(self):
         from multiprocessing import Pool
         traj = pt.iterload('data/tz2.nc', 'data/tz2.parm7')
         for _ in range(10):
-            traj.load(traj.filelist)
+            traj._load(traj.filelist)
         saved_angle = pt.angle(traj, ':3 :10 :11')
         saved_dist = pt.distance(traj, ':3 :10')
 
-        for n_cores in [2, 3, 4, 5, 6, 7, 8]:
+        for n_cores in [2, 3]:
             lines = ['angle :3 :10 :11', 'distance :3 :10']
             pfuncs = partial(_worker_state,
                              n_cores=n_cores,
@@ -60,8 +59,8 @@ class TestWorkerState(unittest.TestCase):
                                      for data in sorted(data_list,
                                                         key=lambda x: x[0]))
             final_data = concat_dict(data_list_sorted_rank)
-            aa_eq(final_data['Ang_00001'], saved_angle)
-            aa_eq(final_data['Dis_00002'], saved_dist)
+            aa_eq(final_data['Ang_00002'], saved_angle)
+            aa_eq(final_data['Dis_00003'], saved_dist)
 
 
 if __name__ == "__main__":
